Update to v5.4.296

+++ src/display/annotation_storage.js
@@ -13,7 +13,7 @@
  * limitations under the License.
  */
 
-import { objectFromMap, unreachable } from "../shared/util.js";
+import { shadow, unreachable } from "../shared/util.js";
 import { AnnotationEditor } from "./editor/editor.js";
 import { MurmurHash3_64 } from "../shared/murmurhash3.js";
 
@@ -29,6 +29,10 @@ const SerializableEmpty = Object.freeze({
 class AnnotationStorage {
   #modified = false;
 
+  #modifiedIds = null;
+
+  #editorsMap = null;
+
   #storage = new Map();
 
   constructor() {
@@ -39,6 +43,17 @@ class AnnotationStorage {
     this.onSetModified = null;
     this.onResetModified = null;
     this.onAnnotationEditor = null;
+
+    if (typeof PDFJSDev === "undefined" || PDFJSDev.test("TESTING")) {
+      // For testing purposes.
+      Object.defineProperty(this, "_setValues", {
+        value: obj => {
+          for (const [key, val] of Object.entries(obj)) {
+            this.setValue(key, val);
+          }
+        },
+      });
+    }
   }
 
   /**
@@ -70,6 +85,13 @@ class AnnotationStorage {
    * @param {string} key
    */
   remove(key) {
+    const storedValue = this.#storage.get(key);
+    if (storedValue === undefined) {
+      return;
+    }
+    if (storedValue instanceof AnnotationEditor) {
+      this.#editorsMap.delete(storedValue.annotationElementId);
+    }
     this.#storage.delete(key);
 
     if (this.#storage.size === 0) {
@@ -109,11 +131,11 @@ class AnnotationStorage {
       this.#setModified();
     }
 
-    if (
-      value instanceof AnnotationEditor &&
-      typeof this.onAnnotationEditor === "function"
-    ) {
-      this.onAnnotationEditor(value.constructor._type);
+    if (value instanceof AnnotationEditor) {
+      (this.#editorsMap ||= new Map()).set(value.annotationElementId, value);
+      if (typeof this.onAnnotationEditor === "function") {
+        this.onAnnotationEditor(value.constructor._type);
+      }
     }
   }
 
@@ -124,22 +146,6 @@ class AnnotationStorage {
    */
   has(key) {
     return this.#storage.has(key);
-  }
-
-  /**
-   * @returns {Object | null}
-   */
-  getAll() {
-    return this.#storage.size > 0 ? objectFromMap(this.#storage) : null;
-  }
-
-  /**
-   * @param {Object} obj
-   */
-  setAll(obj) {
-    for (const [key, val] of Object.entries(obj)) {
-      this.setValue(key, val);
-    }
   }
 
   get size() {
@@ -216,9 +222,23 @@ class AnnotationStorage {
   get editorStats() {
     let stats = null;
     const typeToEditor = new Map();
+    let numberOfEditedComments = 0;
+    let numberOfDeletedComments = 0;
     for (const value of this.#storage.values()) {
       if (!(value instanceof AnnotationEditor)) {
+        if (value.popup) {
+          if (value.popup.deleted) {
+            numberOfDeletedComments += 1;
+          } else {
+            numberOfEditedComments += 1;
+          }
+        }
         continue;
+      }
+      if (value.isCommentDeleted) {
+        numberOfDeletedComments += 1;
+      } else if (value.hasEditedComment) {
+        numberOfEditedComments += 1;
       }
       const editorStats = value.telemetryFinalData;
       if (!editorStats) {
@@ -243,10 +263,63 @@ class AnnotationStorage {
         counters.set(val, count + 1);
       }
     }
+    if (numberOfDeletedComments > 0 || numberOfEditedComments > 0) {
+      stats ||= Object.create(null);
+      stats.comments = {
+        deleted: numberOfDeletedComments,
+        edited: numberOfEditedComments,
+      };
+    }
+    if (!stats) {
+      return null;
+    }
     for (const [type, editor] of typeToEditor) {
       stats[type] = editor.computeTelemetryFinalData(stats[type]);
     }
     return stats;
+  }
+
+  resetModifiedIds() {
+    this.#modifiedIds = null;
+  }
+
+  updateEditor(annotationId, data) {
+    const value = this.#editorsMap?.get(annotationId);
+    if (value) {
+      value.updateFromAnnotationLayer(data);
+      return true;
+    }
+    return false;
+  }
+
+  getEditor(annotationId) {
+    return this.#editorsMap?.get(annotationId) || null;
+  }
+
+  /**
+   * @returns {{ids: Set<string>, hash: string}}
+   */
+  get modifiedIds() {
+    if (this.#modifiedIds) {
+      return this.#modifiedIds;
+    }
+    const ids = [];
+    if (this.#editorsMap) {
+      for (const value of this.#editorsMap.values()) {
+        if (!value.serialize()) {
+          continue;
+        }
+        ids.push(value.annotationElementId);
+      }
+    }
+    return (this.#modifiedIds = {
+      ids: new Set(ids),
+      hash: ids.join(","),
+    });
+  }
+
+  [Symbol.iterator]() {
+    return this.#storage.entries();
   }
 }
 
@@ -281,6 +354,13 @@ class PrintAnnotationStorage extends AnnotationStorage {
    */
   get serializable() {
     return this.#serializable;
+  }
+
+  get modifiedIds() {
+    return shadow(this, "modifiedIds", {
+      ids: new Set(),
+      hash: "",
+    });
   }
 }
